+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Status for the PCILoader (`initial`, `loading`, `loaded`, `error`)
+
 ## v1.0.2 [2025-08-26]
 
 ## Fixed